refactor: move routes to navigation package
