+++ app/src/main/java/com/fpf/smartscan/navigation/Routes.kt
@@ -1,4 +1,4 @@
-package com.fpf.smartscan.constants
+package com.fpf.smartscan.navigation
 
 object Routes {
     const val SEARCH = "search"
@@ -8,21 +8,8 @@ object Routes {
     const val SETTINGS = "settings"
     const val SETTINGS_DETAIL = "settings_detail/{type}"
     const val DONATE = "donate"
-    const val HELP = "help"
     fun settingsDetail(type: String) = "settings_detail/$type"
     fun viewCollection(collectionName: String, clusterId: Long? = null): String {
         val clusterId = clusterId?: -1L
         return "collection_items/$collectionName/$clusterId"
     }}
-
-object SettingTypes {
-    const val THRESHOLD = "threshold"
-    const val MODELS = "models"
-    const val MANAGE_MODELS = "manage_models"
-    const val SEARCHABLE_IMG_DIRS = "searchable_img_dirs"
-    const val SEARCHABLE_VID_DIRS = "searchable_vid_dirs"
-    const val BACKUP_RESTORE = "backup_restore"
-
-    const val DUPLICATES = "duplicates"
-
-}
